starter_code/requirements.txt
Flask_SQLAlchemy==2.5.1
SQLAlchemy==1.4.7
Flask_Moment==1.0.1
Flask_Migrate==3.0.1
Flask_WTF==0.15.1
WTForms==2.3.3
alembic==1.6.5
Flask==1.1.2
Babel==2.9.0
Fabric==2.6.0
python_dateutil==2.8.1

